+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/metrics/MonitoringConfiguration.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,10 +17,12 @@ package com.hotels.bdp.waggledance.metrics;
 
 import java.time.Duration;
 
+import org.springframework.boot.actuate.autoconfigure.metrics.MeterRegistryCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import io.micrometer.core.instrument.Clock;
+import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.config.NamingConvention;
 import io.micrometer.core.instrument.util.HierarchicalNameMapper;
 import io.micrometer.graphite.GraphiteConfig;
@@ -28,8 +30,10 @@ import io.micrometer.graphite.GraphiteMeterRegistry;
 import io.micrometer.graphite.GraphiteProtocol;
 import io.micrometer.jmx.JmxConfig;
 import io.micrometer.jmx.JmxMeterRegistry;
+import io.micrometer.prometheus.PrometheusMeterRegistry;
 
 import com.hotels.bdp.waggledance.conf.GraphiteConfiguration;
+import com.hotels.bdp.waggledance.conf.PrometheusConfiguration;
 
 @Configuration
 public class MonitoringConfiguration {
@@ -111,6 +115,15 @@ public class MonitoringConfiguration {
   @Bean
   public JmxMeterRegistry jmxMeterRegistry() {
     return new JmxMeterRegistry(JmxConfig.DEFAULT, Clock.SYSTEM);
+  }
+
+  @Bean
+  MeterRegistryCustomizer<MeterRegistry> metricsCommonTags(PrometheusConfiguration prometheusConfiguration) {
+    return registry -> {
+      if (registry instanceof PrometheusMeterRegistry) {
+        registry.config().commonTags("application", prometheusConfiguration.getPrefix());
+      }
+    };
   }
 
 }